+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/item/block/ItemStorageBlockBlockItem.java
@@ -48,16 +48,14 @@ public class ItemStorageBlockBlockItem extends AbstractStorageContainerBlockItem
             return;
         }
         final StorageRepository storageRepository = PlatformApi.INSTANCE.getStorageRepository(level);
-        final boolean showCapacityAndProgress = variant != ItemStorageType.Variant.CREATIVE;
+        final boolean hasCapacity = variant != ItemStorageType.Variant.CREATIVE;
         helper.appendToTooltip(
             stack,
             storageRepository,
             tooltip,
             context,
-            AmountFormatting::formatWithUnits,
             AmountFormatting::format,
-            showCapacityAndProgress,
-            true
+            hasCapacity
         );
     }